+++ apps/client/src/components/auth/AuthorizationHub.js
@@ -9,7 +9,10 @@ import AnimatedBGmini from "@components/auth/animations/AnimatedBGmini.js";
 
 import PasswordInput from "@components/auth/elements/PasswordInput";
 import UserNameInput from "@components/auth/elements/UserNameInput";
+import EmailInput from "@components/auth/elements/EmailInput";
 import ConfirmButton from "@components/auth/elements/ConfirmButton";
+
+import ResetPasswordModal from "@components/modals/ResetPasswordModal.js";
 
 import { getIsMobileView } from "@store/values/IsMobileView.js";
 
@@ -19,10 +22,10 @@ import Medium from "@icons/socials/MediumIcon.svg?react";
 import GitHub from "@icons/socials/GitHubIcon.png";
 
 export default function AuthorizationHub({ showDemoMessage = false }) {
+  const [content, setContent] = useState({});
   const [page, setPage] = useState(
     localStorage.getItem("isUsedBefore") ? "login" : "signup"
   );
-  const [content, setContent] = useState({});
 
   const [scope, animate] = useAnimate();
   const [triggerBGAnimation, setTriggerBGAnimation] = useState(false);
@@ -31,6 +34,9 @@ export default function AuthorizationHub({ showDemoMessage = false }) {
 
   const isLoginPage = page === "login";
 
+  const [isModalOpen, setIsModalOpen] = useState(false);
+
+  useEffect(() => setIsModalOpen(!!localStorage.getItem("reset_email")), []);
   useEffect(() => setTriggerBGAnimation(false));
   const triggerExitAnimation = () => {
     animate([
@@ -89,7 +95,7 @@ export default function AuthorizationHub({ showDemoMessage = false }) {
               {isLoginPage ? "Don`t have account?" : "Already have an account?"}
               &nbsp;
               <span
-                className="text-(--color-accent-dark) !font-normal cursor-pointer hover:text-(--color-accent-light-50) transition-colors duration-200"
+                className="text-accent-dark !font-normal cursor-pointer hover:text-(--color-accent-light-50) transition-colors duration-200"
                 onClick={() => setPage(isLoginPage ? "signup" : "login")}
               >
                 {isLoginPage ? "Sign up" : "Log in"}
@@ -108,12 +114,27 @@ export default function AuthorizationHub({ showDemoMessage = false }) {
             </div>
           ) : null}
           <div className="sm:mt-[25px] flex-1 flex flex-col gap-[15px]">
-            <UserNameInput setState={setContent} />
+            <UserNameInput
+              setState={setContent}
+              isResetModalOpen={isModalOpen}
+            />
+            {isLoginPage ? null : <EmailInput setState={setContent} />}
             <PasswordInput setState={setContent} />
+            {isLoginPage ? (
+              <div className="flex gap-[7px]">
+                <span
+                  onClick={() => setIsModalOpen(true)}
+                  className="text-accent-dark cursor-pointer !font-normal"
+                >
+                  Forgot password?
+                </span>
+              </div>
+            ) : null}
             <ConfirmButton
               page={page}
               content={content}
               onClickEvent={triggerExitAnimation}
+              isResetModalOpen={isModalOpen}
             />
           </div>
           <div className="w-full flex flex-row items-center gap-[10px]">
@@ -203,6 +224,10 @@ export default function AuthorizationHub({ showDemoMessage = false }) {
           </div>
         </div>
       </m.div>
+      <ResetPasswordModal
+        isOpen={isModalOpen}
+        onClose={() => setIsModalOpen(false)}
+      />
     </section>
   );
 }