+++ apps/client/src/components/context/elements/MessageActions.js
@@ -3,8 +3,11 @@ import { useSelector, useDispatch } from "react-redux";
 import { useLocation } from "react-router";
 
 import draftService from "@services/tools/draftService.js";
+import messagesService from "@services/messagesService.js";
 
 import ContextLink from "@components/context/elements/ContextLink";
+
+import { useConfirmWindow } from "@hooks/useConfirmWindow.js";
 
 import { addExternalProps } from "@store/values/ContextMenu.js";
 import { getSelectedConversationId } from "@store/values/SelectedConversation.js";
@@ -15,14 +18,18 @@ import upsertMidsInPath from "@utils/navigation/upasert_mids_in_path.js";
 import writeToCanvas from "@utils/media/write_to_canvas.js";
 
 import Reply from "@icons/context/Reply.svg?react";
+import Edit from "@icons/context/EditText.svg?react";
 import Copy from "@icons/context/Copy.svg?react";
 import Forward from "@icons/context/Forward.svg?react";
 import Select from "@icons/context/Select.svg?react";
 import Download from "@icons/context/Download.svg?react";
+import Delete from "@icons/context/Delete.svg?react";
 
 export default function MessageActions({ listOfIds }) {
   const dispatch = useDispatch();
   const location = useLocation();
+
+  const confirmWindow = useConfirmWindow();
 
   const selectedCID = useSelector(getSelectedConversationId);
 
@@ -124,6 +131,39 @@ export default function MessageActions({ listOfIds }) {
             location.pathname + location.hash,
             `/forward?mids=[${message._id}]`
           );
+        }}
+      />
+    ),
+    messageEdit: (
+      <ContextLink
+        key={"messageEdit"}
+        text="Edit"
+        icon={<Edit />}
+        onClick={() => {
+          dispatch(
+            addExternalProps({
+              [selectedCID]: { draft_edited_mid: message._id },
+            })
+          );
+          draftService.saveDraft(selectedCID, { edited_mid: message._id });
+        }}
+      />
+    ),
+    messageDelete: (
+      <ContextLink
+        key={"messageDelete"}
+        text="Delete"
+        icon={<Delete />}
+        onClick={async () => {
+          const { isConfirm, data } = await confirmWindow({
+            title: "Delete selected message?",
+            confirmText: "Delete",
+            cancelText: "Cancel",
+            action: "messageDelete",
+          });
+          const { _id } = message;
+          isConfirm &&
+            messagesService.sendMessageDelete(selectedCID, [_id], data.type);
         }}
       />
     ),