+++ apps/client/src/components/hub/chatForm/ChatFormContent.js
@@ -1,3 +1,4 @@
+import { AnimatePresence } from "motion/react";
 import { useMemo, useRef } from "react";
 import { useSelector } from "react-redux";
 
@@ -7,6 +8,7 @@ import ChatFormInput from "@components/hub/chatForm/ChatFormInput.js";
 import ChatFormInputContent from "@components/hub/chatForm/ChatFormInputContent.js";
 import CustomScrollBar from "@components/_helpers/CustomScrollBar";
 import MessagesList from "@components/hub/elements/MessagesList";
+import SummaryContainer from "@components/hub/elements/SummaryContainer.js";
 
 import { getConverastionById } from "@store/values/Conversations.js";
 import { selectMessagesEntities } from "@store/values/Messages.js";
@@ -23,6 +25,7 @@ export default function ChatFormContent() {
   const messages = Object.values(messagesEntities);
 
   const draftExtenralProps = useSelector(selectContextExternalProps);
+
   const draftRepliedMessage = useMemo(() => {
     const repliedMessageId =
       draftExtenralProps[selectedCID]?.draft_replied_mid ||
@@ -32,6 +35,12 @@ export default function ChatFormContent() {
   const draftForwardedMessage = useMemo(() => {
     const forwardedMessageId = selectedConversation?.draft?.forwarded_mids;
     return forwardedMessageId?.map((mid) => messagesEntities[mid]);
+  }, [selectedConversation, draftExtenralProps, messagesEntities]);
+  const draftEditedMessage = useMemo(() => {
+    const editedMessageId =
+      draftExtenralProps[selectedCID]?.draft_edited_mid ||
+      draftService.getDraftEditedMessageId(selectedCID);
+    return messagesEntities[editedMessageId];
   }, [selectedConversation, draftExtenralProps, messagesEntities]);
 
   const chatContentView = useMemo(() => {
@@ -58,12 +67,19 @@ export default function ChatFormContent() {
 
   return (
     <>
+      <AnimatePresence>
+        <SummaryContainer summaryContent={selectedConversation?.summary} />
+      </AnimatePresence>
       {chatContentView}
       <ChatFormInputContent
+        editedMessage={draftEditedMessage}
         repliedMessage={draftRepliedMessage}
         forwardedMessages={draftForwardedMessage}
       />
-      <ChatFormInput chatMessagesBlockRef={chatMessagesBlock} />
+      <ChatFormInput
+        chatMessagesBlockRef={chatMessagesBlock}
+        editedMessage={draftEditedMessage}
+      />
     </>
   );
 }