+++ apps/client/src/components/hub/elements/MessagesList.js
@@ -18,6 +18,7 @@ import {
 } from "@store/values/Participants";
 import { getConverastionById } from "@store/values/Conversations";
 import { selectCurrentUserId } from "@store/values/CurrentUserId";
+import { getIsMobileView } from "@store/values/IsMobileView.js";
 
 import upsertMidsInPath from "@utils/navigation/upasert_mids_in_path.js";
 
@@ -26,6 +27,8 @@ import ArrowBottom from "@icons/options/ArrowBottom.svg?react";
 export default function MessagesList({ scrollRef: scrollableContainer }) {
   const dispatch = useDispatch();
   const { pathname, hash } = useLocation();
+
+  const isMobile = useSelector(getIsMobileView);
 
   const timer = useRef(null);
 
@@ -353,6 +356,7 @@ export default function MessagesList({ scrollRef: scrollableContainer }) {
           repliedMessage={repliedMessage}
           sender={participants[from]}
           currentUserId={currentUserId}
+          isMobile={isMobile}
           isSelected={isSelected}
           isSelectionMode={isSelectionMode}
           isPrevMesssageYours={isPrevMesssageYours}