+++ apps/client/src/services/autoLoginService.js
@@ -108,6 +108,53 @@ class AutoLoginService {
       showCustomAlert(error.message, "warning");
     }
   }
+
+  async sendOtpToken(email) {
+    try {
+      await api.userSendOTPToken({ email });
+      localStorage.setItem("reset_email", email);
+    } catch (err) {
+      showCustomAlert(
+        err.message || "We couldn’t find an account with this email.",
+        "warning"
+      );
+      return false;
+    }
+    return true;
+  }
+
+  async resendOtpToken(email) {
+    try {
+      await api.userSendOTPToken({ email });
+      showCustomAlert("OTP sent.", "success");
+    } catch (err) {
+      showCustomAlert(
+        err.message || "Failed to resend token. Try again.",
+        "warning"
+      );
+      return false;
+    }
+    return true;
+  }
+
+  async sendResetPassword(email, token, newPassword) {
+    try {
+      await api.userResetPassword({
+        email,
+        token: +token,
+        new_password: newPassword,
+      });
+      localStorage.removeItem("reset_email");
+      showCustomAlert("Password successfully changed.", "success");
+    } catch (err) {
+      showCustomAlert(
+        err.message || "Failed to reset password. Try again.",
+        "warning"
+      );
+      return false;
+    }
+    return true;
+  }
 }
 
 const autoLoginService = new AutoLoginService();